:b: yeah
not really sure what I need to do next to get this to work. I'm tempted to start over and try to make the game seperately, then port it over to discord after the fact. Perhaps like have a class with a listener that calls methods w/ a central main method? dunno tbh.